requirements.dev.txt
# This file is deliberately not locked with pip-tools so as to allow
# each CI target platform's pip to find suitable versions itself.
pytest-cov~=4.0
pytest~=7.0
requests-mock~=1.10